Ares v5 (#294)
* update from dev

* cleaning

* ROUND3: last minute update

* bow fixing and exploring map - to be checked

* refactoring part 1

* Refactoring done: strategies

---------
